backend/requirements.txt
django
django-ninja
django-phonenumber-field[phonenumbers]
python-dotenv
requests
python-jose
channels
channels-redis
daphne
whitenoise
redis
